Build, Draw, and Save Identicon
After filtering out the odd squares, build a pixel map and include it
in the Identicon.Image struct by calculating the start (top left
corner) and stop  (bottom right corner) points of a square on the grid.
Include the `:egd` Erlang library for actual drawing and use it to
create a base grid with an assigned fill color. Iterate over each
square in the pixel map to fill it, then render the image. Finally,
write the image to a file using the originally input string as the
filename